// src/gen/api/types/FriendRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { UserSummary } from './UserSummary.ts'

export type FriendRequest = {
  /**
   * @type object
   */
  user: UserSummary
  /**
   * @description UUID identifier
   * @type string, uuid
   */
  requesterId: string
  /**
   * @description ISO 8601 timestamp
   * @type string, date-time
   */
  createdAt: string
}